% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadCausalPathFormat.R
\name{loadCausalPathFormat}
\alias{loadCausalPathFormat}
\title{Generate igraph Graph with node, edge colors and highlighted nodes.}
\usage{
loadCausalPathFormat(sif, formatFile, highlightNodes = NULL,
  formatConfFile = NULL)
}
\arguments{
\item{sif}{Network which is loaded in graph and gets colors assigned}

\item{formatFile}{File path of the .format file with information about the colors of the nodes}

\item{highlightNodes}{List of namesead  of nodes which are highlighted by a circle node instead of rectangles}

\item{formatConfFile}{File path of .format file of conflicting network to plot merged network. Required for
color of nodes which are only present in subnetwork which was added from the conflicting network.}
}
\value{
igraph graph object with adapted colors
}
\description{
Generate igraph Graph with node, edge colors and highlighted nodes.
}
\concept{
rcausalpath
}
